rename datastacks_cli
